+++ src/main/java/dev/yatloaf/modkrowd/config/Config.java
@@ -12,7 +12,9 @@ import dev.yatloaf.modkrowd.config.exception.ReadConfigException;
 import dev.yatloaf.modkrowd.config.exception.UnsupportedVersionConfigException;
 import dev.yatloaf.modkrowd.config.exception.WriteConfigException;
 import dev.yatloaf.modkrowd.config.feature.Feature;
+import dev.yatloaf.modkrowd.util.Util;
 import net.fabricmc.loader.api.SemanticVersion;
+import net.fabricmc.loader.api.Version;
 import net.fabricmc.loader.api.VersionParsingException;
 import net.minecraft.util.GsonHelper;
 
@@ -28,14 +30,21 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class Config {
-    private final Map<Feature, FeatureState> states = new HashMap<>();
+    public static final Version COMPATIBILITY_VERSION = Util.assertSuccess(() -> SemanticVersion.parse("0.3.0"));
 
+    protected final Map<Feature, FeatureState> states = new HashMap<>();
+    protected final Map<String, JsonElement> unknownStates = new HashMap<>();
+
+    public Version version;
+    public Version compatibilityVersion;
     public FeatureTab selectedTab;
 
     public Config() {
         for (Feature feature : Features.FEATURES) {
             this.states.put(feature, feature.makeState());
         }
+        this.version = ModKrowd.VERSION;
+        this.compatibilityVersion = COMPATIBILITY_VERSION;
         this.selectedTab = Features.APPEARANCE;
     }
 
@@ -45,6 +54,9 @@ public class Config {
             featureState.imitate(source.getState(feature));
             this.states.put(feature, featureState);
         }
+        this.unknownStates.putAll(source.unknownStates);
+        this.version = source.version;
+        this.compatibilityVersion = source.compatibilityVersion;
         this.selectedTab = source.selectedTab;
     }
 
@@ -53,13 +65,27 @@ public class Config {
 
             JsonObject root = JsonParser.parseReader(jsonReader).getAsJsonObject();
 
+            boolean compatibilityMode;
             SemanticVersion version = getSemanticVersion(root, "version");
             if (version.compareTo(ModKrowd.VERSION) > 0) {
-                throw new UnsupportedVersionConfigException(
-                        "Expected version %s or lower, found %s".formatted(ModKrowd.VERSION.getFriendlyString(), version.getFriendlyString())
-                );
+                this.version = version;
+                this.compatibilityVersion = getSemanticVersion(root, "compatibility_version");
+                if (this.compatibilityVersion.compareTo(COMPATIBILITY_VERSION) > 0) {
+                    throw new UnsupportedVersionConfigException(
+                            "Expected compatibility version %s or lower, found %s"
+                                    .formatted(COMPATIBILITY_VERSION.getFriendlyString(), this.compatibilityVersion.getFriendlyString())
+                    );
+                } else {
+                    ModKrowd.LOGGER.info("[ModKrowd] Config from version {} found, using {} compatibility mode", version, this.compatibilityVersion);
+                    compatibilityMode = true;
+                }
+            } else {
+                this.version = ModKrowd.VERSION;
+                this.compatibilityVersion = COMPATIBILITY_VERSION;
+                compatibilityMode = false;
             }
 
+            this.unknownStates.clear();
             this.selectedTab = Features.ID_TO_TAB.getOrDefault(GsonHelper.getAsString(root, "selected_tab", ""), Features.APPEARANCE);
 
             if (GsonHelper.isObjectNode(root, "features")) {
@@ -78,6 +104,8 @@ public class Config {
                         } else {
                             state.deserialize(GsonHelper.convertToJsonObject(value, key));
                         }
+                    } else if (compatibilityMode) {
+                        this.unknownStates.put(key, value);
                     }
                 }
             }
@@ -96,7 +124,8 @@ public class Config {
             jsonWriter.setIndent("    ");
 
             jsonWriter.beginObject();
-                jsonWriter.name("version").value(ModKrowd.VERSION.getFriendlyString());
+                jsonWriter.name("version").value(this.version.getFriendlyString());
+                jsonWriter.name("compatibility_version").value(this.compatibilityVersion.getFriendlyString());
                 jsonWriter.name("selected_tab").value(this.selectedTab.id);
                 jsonWriter.name("features").beginObject();
                     for (Feature feature : Features.FEATURES) {
@@ -104,6 +133,9 @@ public class Config {
                         JsonObject dest = new JsonObject();
                         this.getState(feature).serialize(dest);
                         jsonWriter.jsonValue(dest.toString());
+                    }
+                    for (Map.Entry<String, JsonElement> entry : this.unknownStates.entrySet()) {
+                        jsonWriter.name(entry.getKey()).jsonValue(entry.getValue().toString());
                     }
                 jsonWriter.endObject();
             jsonWriter.endObject();
@@ -117,6 +149,9 @@ public class Config {
         for (FeatureState featureState : this.states()) {
             featureState.enabled = false;
         }
+        this.unknownStates.clear();
+        this.version = ModKrowd.VERSION;
+        this.compatibilityVersion = COMPATIBILITY_VERSION;
     }
 
     public FeatureState getState(Feature feature) {